+++ tests/Feature/DefaultConfigTest.php
@@ -1,7 +1,9 @@
 <?php
 
+use TailwindClassMerge\Support\Config;
+
 test('default config has correct types', function () {
-    $defaultConfig = \TailwindClassMerge\Support\Config::getDefaultConfig();
+    $defaultConfig = Config::getDefaultConfig();
 
     expect($defaultConfig)
         ->not->toHaveKey('nonExistent')