+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## v3.3.1 (2026-03-18)
+### Fixed
+-   Fix PHPStan type issue in class list normalization for `Arr::toCssClasses`
+-   Remove redundant null coalescing in arbitrary value validator regex handling
+### Updated
+-   Update package dependencies
+
 ## v3.3.0 (2026-02-04)
 ### Updated
 -   Update package dependencies